cookies/http-state/0001.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
  async_test(t => {
    window.addEventListener("message", t.step_func_done(e => {
      assert_equals(e.data.cookie, "foo=bar");
    }));

    let i = document.createElement('iframe');
    i.src = "./resources/cookie-setter.py?file=0001";
    document.body.appendChild(i);    
  });
</script>
